Added a hardcoded string category collection
This is because the category and name pairs in the *.tdbl is not up to
date with the actual category + rdbIds. Thus, there can be more category,
rdbId pairs than there are category and name pairs. Since we do not want
to rely on id's to glance over the file names, we give them hardcoded
canonical names